Moved section number parsing out of Section.add_meetings() to eliminate a bug and reduce redundant calls. Also improved Section.__lt__() and added a TODO.
By moving the calls to Section.fix_section_number() to the end of the places it is initialized (Section.__init__() and Course.add_section()), a bug in which section numbers were not shortened properly has been eliminated.

Also, improvements to Section.__lt__() now prevent comparison of a section without a number.

TODO: Allow the Section class to store information for classes with both labs and conferences.